requirements.txt
# --------------------------------------------------------------------------------------
# Core Dependencies
# --------------------------------------------------------------------------------------
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# --------------------------------------------------------------------------------------
# OpenAI and LLM Dependencies
# --------------------------------------------------------------------------------------
openai==1.55.3
tiktoken==0.8.0
anthropic>=0.21.0
google-generativeai>=0.3.0

# --------------------------------------------------------------------------------------
# Vector and Graph Dependencies
# --------------------------------------------------------------------------------------
numpy==1.26.4
faiss-cpu==1.9.0
networkx==3.4.2

# --------------------------------------------------------------------------------------
# Web and CORS
# --------------------------------------------------------------------------------------
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.7.0
